search-api/pyproject.toml
[project]
name = "search-api"
version = "0.1.0"
description = "Search API for RAG Scrapbox"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "elasticsearch>=8.0.0,<9.0.0",
    "aiohttp>=3.8.0",
    "httpx>=0.24.0",
    "google-genai>=0.1.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "tenacity>=8.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B"]

[tool.ruff.isort]
known-first-party = ["api"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.uv]
package = false
